server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1

# AI optional (install if using AI endpoints)
# ultralytics==8.3.59
# opencv-python==4.11.0.86
# onnxruntime==1.19.0
# numpy==2.1.3
